requirements.txt
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
entrypoints==0.3
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
numpy==1.24.4
oauthlib==3.1.0
paho-mqtt==1.6.1
pandas==2.0.3
PyJWT==1.7.1
pymodbus==2.5.3
pyserial==3.5
python-apt==2.0.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
tzdata==2023.3
urllib3==1.25.8
wadllib==1.3.3
